Refactor carousel classes: add a new parent class for both page objects.
Use new classes in a test case.
